update some modification to remove old version codes.
